fix: multiple assignment fixes for macos scripts and cloudpc provisioning policies (#612)
* fix: added assignment removal to deletion of CloudPcProvisioningPolicy resource

* fix: for assignment logic for macos_platform_script and macos_custom_attribute_script
