+++ internal/services/resources/device_management/graph_beta/macos_platform_script/model.go
@@ -13,8 +13,6 @@ type MacOSPlatformScriptResourceModel struct {
 	DisplayName                 types.String                                                `tfsdk:"display_name"`
 	Description                 types.String                                                `tfsdk:"description"`
 	ScriptContent               types.String                                                `tfsdk:"script_content"`
-	CreatedDateTime             types.String                                                `tfsdk:"created_date_time"`
-	LastModifiedDateTime        types.String                                                `tfsdk:"last_modified_date_time"`
 	RunAsAccount                types.String                                                `tfsdk:"run_as_account"`
 	FileName                    types.String                                                `tfsdk:"file_name"`
 	RoleScopeTagIds             types.Set                                                   `tfsdk:"role_scope_tag_ids"`
