+++ internal/services/resources/device_management/graph_beta/macos_platform_script/resource.go
@@ -111,14 +111,6 @@ func (r *MacOSPlatformScriptResource) Schema(ctx context.Context, req resource.S
 				Required:            true,
 				Sensitive:           true,
 			},
-			"created_date_time": schema.StringAttribute{
-				MarkdownDescription: "The date and time the macOS Platform Script was created. This property is read-only.",
-				Computed:            true,
-			},
-			"last_modified_date_time": schema.StringAttribute{
-				MarkdownDescription: "The date and time the macOS Platform Script was last modified. This property is read-only.",
-				Computed:            true,
-			},
 			"run_as_account": schema.StringAttribute{
 				MarkdownDescription: "Indicates the type of execution context. Possible values are: `system`, `user`.",
 				Required:            true,
