+++ internal/services/resources/device_management/graph_beta/macos_platform_script/state.go
@@ -1,4 +1,5 @@
-package graphBetaMacOSCustomAttributeScript
+// MapRemoteResourceStateToTerraform states the base properties of a SettingsCatalogProfileResourceModel to a Terraform state
+package graphBetaMacOSPlatformScript
 
 import (
 	"context"
@@ -12,41 +13,40 @@ import (
 	graphmodels "github.com/microsoftgraph/msgraph-beta-sdk-go/models"
 )
 
-// MapRemoteStateToTerraform maps the remote DeviceCustomAttributeShellScript to the Terraform resource model
-func MapRemoteStateToTerraform(ctx context.Context, data *DeviceCustomAttributeShellScriptResourceModel, remoteResource graphmodels.DeviceCustomAttributeShellScriptable) {
+// MapRemoteResourceStateToTerraform maps the base properties of a MacOSPlatformScriptResourceModel to a Terraform state.
+func MapRemoteResourceStateToTerraform(ctx context.Context, data *MacOSPlatformScriptResourceModel, remoteResource graphmodels.DeviceShellScriptable) {
 	if remoteResource == nil {
 		tflog.Debug(ctx, "Remote resource is nil")
 		return
 	}
 
 	tflog.Debug(ctx, "Starting to map remote state to Terraform state", map[string]interface{}{
-		"resourceId":   remoteResource.GetId(),
 		"resourceName": remoteResource.GetDisplayName(),
+		"resourceId":   remoteResource.GetId(),
 	})
 
 	data.ID = convert.GraphToFrameworkString(remoteResource.GetId())
-	data.CustomAttributeName = convert.GraphToFrameworkString(remoteResource.GetCustomAttributeName())
-	data.CustomAttributeType = convert.GraphToFrameworkEnum(remoteResource.GetCustomAttributeType())
 	data.DisplayName = convert.GraphToFrameworkString(remoteResource.GetDisplayName())
 	data.Description = convert.GraphToFrameworkString(remoteResource.GetDescription())
-	data.ScriptContent = convert.GraphToFrameworkBytes(remoteResource.GetScriptContent())
 	data.RunAsAccount = convert.GraphToFrameworkEnum(remoteResource.GetRunAsAccount())
 	data.FileName = convert.GraphToFrameworkString(remoteResource.GetFileName())
 	data.RoleScopeTagIds = convert.GraphToFrameworkStringSet(ctx, remoteResource.GetRoleScopeTagIds())
-	data.CreatedDateTime = convert.GraphToFrameworkTime(remoteResource.GetCreatedDateTime())
-	data.LastModifiedDateTime = convert.GraphToFrameworkTime(remoteResource.GetLastModifiedDateTime())
+	data.BlockExecutionNotifications = convert.GraphToFrameworkBool(remoteResource.GetBlockExecutionNotifications())
+	data.ExecutionFrequency = convert.GraphToFrameworkISODuration(remoteResource.GetExecutionFrequency())
+	data.ScriptContent = convert.GraphToFrameworkBytes(remoteResource.GetScriptContent())
+	data.RetryCount = convert.GraphToFrameworkInt32(remoteResource.GetRetryCount())
 
 	if assignments := remoteResource.GetAssignments(); len(assignments) > 0 {
 		MapAssignmentsToTerraform(ctx, data, assignments)
 	}
 
-	tflog.Debug(ctx, fmt.Sprintf("Finished mapping remote state for resource %s with id %s", ResourceName, data.ID.ValueString()))
+	tflog.Debug(ctx, fmt.Sprintf("Finished stating resource %s with id %s", ResourceName, data.ID.ValueString()))
 }
 
 // MapAssignmentsToTerraform processes script assignments directly from the slice returned by GetAssignments
 // There appears to be no other way to do this, as the assignments are not returned by any other api call
 // despite all of the docs saying there is.
-func MapAssignmentsToTerraform(ctx context.Context, data *DeviceCustomAttributeShellScriptResourceModel, assignments []graphmodels.DeviceManagementScriptAssignmentable) {
+func MapAssignmentsToTerraform(ctx context.Context, data *MacOSPlatformScriptResourceModel, assignments []graphmodels.DeviceManagementScriptAssignmentable) {
 	if len(assignments) > 0 {
 		tflog.Debug(ctx, "No assignments to process")
 		return
@@ -60,7 +60,7 @@ func MapAssignmentsToTerraform(ctx context.Context, data *DeviceCustomAttributeS
 
 // processAssignments handles the direct processing of assignment slices
 // This contains the core logic from MapRemoteAssignmentStateToTerraform but works with the slice type
-func processAssignments(ctx context.Context, data *DeviceCustomAttributeShellScriptResourceModel, assignments []graphmodels.DeviceManagementScriptAssignmentable) {
+func processAssignments(ctx context.Context, data *MacOSPlatformScriptResourceModel, assignments []graphmodels.DeviceManagementScriptAssignmentable) {
 	tflog.Debug(ctx, "Starting to map assignments directly to Terraform state")
 
 	scriptAssignments := &sharedmodels.DeviceManagementScriptAssignmentResourceModel{
